+++ fv3gfs/util/zarr_monitor.py
@@ -137,7 +137,7 @@ class _ZarrVariableWriter:
         self.sync_array()
 
     def _init_zarr_root(self, quantity):
-        tile_shape = self._partitioner.tile.tile_extent(quantity.metadata)
+        tile_shape = self._partitioner.tile.global_extent(quantity.metadata)
         chunks = self._prepend_chunks + array_chunks(
             self._partitioner.layout, tile_shape, quantity.dims
         )
@@ -161,7 +161,7 @@ class _ZarrVariableWriter:
         if self.i_time >= self.array.shape[0] and self.rank == 0:
             new_shape = list(
                 self._prepend_shape
-                + self._partitioner.tile.tile_extent(quantity.metadata)
+                + self._partitioner.tile.global_extent(quantity.metadata)
             )
             new_shape[0] = self.i_time + 1
             self.array.resize(*new_shape)