+++ tests/test_cube_scatter_gather.py
@@ -58,6 +58,14 @@ def time():
     return datetime.datetime(2000, 1, 1)
 
 
+def assert_quantity_equals(result, reference):
+    assert result.dims == reference.dims
+    assert result.units == reference.units
+    assert result.extent == reference.extent
+    assert isinstance(result.data, type(reference.data))
+    reference.np.testing.assert_array_equal(result.view[:], reference.view[:])
+
+
 @pytest.fixture()
 def dim_lengths(layout):
     return {
@@ -72,14 +80,14 @@ def dim_lengths(layout):
 
 @pytest.fixture()
 def communicator_list(layout):
-    total_ranks = layout[0] * layout[1]
+    total_ranks = 6 * layout[0] * layout[1]
     shared_buffer = {}
     return_list = []
     for rank in range(total_ranks):
         return_list.append(
-            fv3gfs.util.TileCommunicator(
+            fv3gfs.util.CubedSphereCommunicator(
                 fv3gfs.util.testing.DummyComm(rank, total_ranks, shared_buffer),
-                fv3gfs.util.TilePartitioner(layout),
+                fv3gfs.util.CubedSpherePartitioner(fv3gfs.util.TilePartitioner(layout)),
             )
         )
     return return_list
@@ -94,38 +102,40 @@ def tile_extent(dims, dim_lengths):
 
 
 @pytest.fixture
-def tile_quantity(dims, units, dim_lengths, tile_extent, n_tile_halo, numpy):
-    return get_tile_quantity(dims, units, dim_lengths, tile_extent, n_tile_halo, numpy)
+def cube_quantity(dims, units, dim_lengths, tile_extent, n_tile_halo, numpy):
+    return get_cube_quantity(dims, units, dim_lengths, tile_extent, n_tile_halo, numpy)
 
 
 @pytest.fixture
-def scattered_quantities(tile_quantity, layout, n_rank_halo, numpy):
+def scattered_quantities(cube_quantity, layout, n_rank_halo, numpy):
+    tile_ranks = layout[0] * layout[1]
     return_list = []
-    total_ranks = layout[0] * layout[1]
     partitioner = fv3gfs.util.TilePartitioner(layout)
-    for rank in range(total_ranks):
-        # partitioner is tested in other tests, here we assume it works
-        subtile_slice = partitioner.subtile_slice(
-            rank,
-            global_dims=tile_quantity.dims,
-            global_extent=tile_quantity.extent,
-            overlap=True,
-        )
-        subtile_view = tile_quantity.view[subtile_slice]
-        subtile_quantity = get_quantity(
-            tile_quantity.dims,
-            tile_quantity.units,
-            subtile_view.shape,
-            n_rank_halo,
-            numpy,
-        )
-        subtile_quantity.view[:] = subtile_view
-        return_list.append(subtile_quantity)
+    for i_tile in range(6):
+        for rank in range(tile_ranks):
+            # partitioner is tested in other tests, here we assume it works
+            subtile_slice = partitioner.subtile_slice(
+                rank,
+                global_dims=cube_quantity.dims[1:],
+                global_extent=cube_quantity.extent[1:],
+                overlap=True,
+            )
+            subtile_view = cube_quantity.view[(i_tile,) + subtile_slice]
+            subtile_quantity = get_quantity(
+                cube_quantity.dims[1:],
+                cube_quantity.units,
+                subtile_view.shape,
+                n_rank_halo,
+                numpy,
+            )
+            subtile_quantity.view[:] = subtile_view
+            return_list.append(subtile_quantity)
     return return_list
 
 
-def get_tile_quantity(dims, units, dim_lengths, tile_extent, n_halo, numpy):
-    extent = [dim_lengths[dim] for dim in dims]
+def get_cube_quantity(dims, units, dim_lengths, tile_extent, n_halo, numpy):
+    extent = [6] + [dim_lengths[dim] for dim in dims]
+    dims = [fv3gfs.util.TILE_DIM] + dims
     quantity = get_quantity(dims, units, extent, n_halo, numpy)
     quantity.view[:] = numpy.random.randn(*quantity.extent)
     return quantity
@@ -146,14 +156,13 @@ def get_quantity(dims, units, extent, n_halo, numpy):
 @pytest.mark.parametrize("backend", ["gt4py_numpy", "gt4py_cupy"], indirect=True)
 @pytest.mark.parametrize("dims, layout", [["x,y,z", (2, 2)]], indirect=True)
 def test_gathered_quantity_has_storage(
-    tile_quantity, scattered_quantities, communicator_list, time, backend
+    scattered_quantities, communicator_list, time, backend
 ):
     for communicator, rank_quantity in reversed(
         list(zip(communicator_list, scattered_quantities))
     ):
         result = communicator.gather(send_quantity=rank_quantity)
         if communicator.rank == 0:
-            print(result.gt4py_backend, result.metadata)
             assert isinstance(result.storage, gt4py.storage.storage.Storage)
         else:
             assert result is None
@@ -162,20 +171,20 @@ def test_gathered_quantity_has_storage(
 @pytest.mark.parametrize("backend", ["gt4py_numpy", "gt4py_cupy"], indirect=True)
 @pytest.mark.parametrize("dims, layout", [["x,y,z", (2, 2)]], indirect=True)
 def test_scattered_quantity_has_storage(
-    tile_quantity, communicator_list, time, backend
+    cube_quantity, communicator_list, time, backend
 ):
     result_list = []
     for communicator in communicator_list:
         if communicator.rank == 0:
-            result_list.append(communicator.scatter(send_quantity=tile_quantity))
+            result_list.append(communicator.scatter(send_quantity=cube_quantity))
         else:
             result_list.append(communicator.scatter())
     for rank, result in enumerate(result_list):
         assert isinstance(result.storage, gt4py.storage.storage.Storage)
 
 
-def test_tile_gather_state(
-    tile_quantity, scattered_quantities, communicator_list, time, backend
+def test_cube_gather_state(
+    cube_quantity, scattered_quantities, communicator_list, time, backend
 ):
     for communicator, rank_quantity in reversed(
         list(zip(communicator_list, scattered_quantities))
@@ -188,17 +197,13 @@ def test_tile_gather_state(
             assert out is None
     assert result_state["time"] == time
     result = result_state["air_temperature"]
-    assert result.dims == tile_quantity.dims
-    assert result.units == tile_quantity.units
-    assert result.extent == tile_quantity.extent
-    assert isinstance(result.data, type(tile_quantity.data))
-    tile_quantity.np.testing.assert_array_equal(result.view[:], tile_quantity.view[:])
+    assert_quantity_equals(result, cube_quantity)
 
 
-def test_tile_gather_state_with_recv_state(
-    tile_quantity, scattered_quantities, communicator_list, time
+def test_cube_gather_state_with_recv_state(
+    cube_quantity, scattered_quantities, communicator_list, time
 ):
-    recv_state = {"time": time, "air_temperature": copy.deepcopy(tile_quantity)}
+    recv_state = {"time": time, "air_temperature": copy.deepcopy(cube_quantity)}
     recv_state["air_temperature"].data[:] = -1
     for communicator, rank_quantity in reversed(
         list(zip(communicator_list, scattered_quantities))
@@ -210,14 +215,11 @@ def test_tile_gather_state_with_recv_state(
             communicator.gather_state(send_state=state)
     assert recv_state["time"] == time
     result = recv_state["air_temperature"]
-    assert result.dims == tile_quantity.dims
-    assert result.units == tile_quantity.units
-    assert result.extent == tile_quantity.extent
-    tile_quantity.np.testing.assert_array_equal(result.view[:], tile_quantity.view[:])
+    assert_quantity_equals(result, cube_quantity)
 
 
-def test_tile_gather_no_recv_quantity(
-    tile_quantity, scattered_quantities, communicator_list
+def test_cube_gather_no_recv_quantity(
+    cube_quantity, scattered_quantities, communicator_list
 ):
     for communicator, rank_quantity in reversed(
         list(zip(communicator_list, scattered_quantities))
@@ -225,30 +227,24 @@ def test_tile_gather_no_recv_quantity(
         result = communicator.gather(send_quantity=rank_quantity)
         if communicator.rank != 0:
             assert result is None
-    assert result.dims == tile_quantity.dims
-    assert result.units == tile_quantity.units
-    assert result.extent == tile_quantity.extent
-    tile_quantity.np.testing.assert_array_equal(result.view[:], tile_quantity.view[:])
+    assert_quantity_equals(result, cube_quantity)
 
 
-def test_tile_scatter_no_recv_quantity(
-    tile_quantity, scattered_quantities, communicator_list
+def test_cube_scatter_no_recv_quantity(
+    cube_quantity, scattered_quantities, communicator_list
 ):
     result_list = []
     for communicator in communicator_list:
         if communicator.rank == 0:
-            result_list.append(communicator.scatter(send_quantity=tile_quantity))
+            result_list.append(communicator.scatter(send_quantity=cube_quantity))
         else:
             result_list.append(communicator.scatter())
     for rank, (result, scattered) in enumerate(zip(result_list, scattered_quantities)):
-        assert result.dims == scattered.dims
-        assert result.units == scattered.units
-        assert result.extent == scattered.extent
-        scattered.np.testing.assert_array_equal(result.view[:], scattered.view[:])
+        assert_quantity_equals(result, scattered)
 
 
-def test_tile_scatter_with_recv_quantity(
-    tile_quantity, scattered_quantities, communicator_list
+def test_cube_scatter_with_recv_quantity(
+    cube_quantity, scattered_quantities, communicator_list
 ):
     recv_quantities = copy.deepcopy(scattered_quantities)
     for q in recv_quantities:
@@ -256,7 +252,7 @@ def test_tile_scatter_with_recv_quantity(
     for recv, communicator in zip(recv_quantities, communicator_list):
         if communicator.rank == 0:
             result = communicator.scatter(
-                send_quantity=tile_quantity, recv_quantity=recv
+                send_quantity=cube_quantity, recv_quantity=recv
             )
         else:
             result = communicator.scatter(recv_quantity=recv)
@@ -264,16 +260,13 @@ def test_tile_scatter_with_recv_quantity(
     for rank, (result, scattered) in enumerate(
         zip(recv_quantities, scattered_quantities)
     ):
-        assert result.dims == scattered.dims
-        assert result.units == scattered.units
-        assert result.extent == scattered.extent
-        scattered.np.testing.assert_array_equal(result.view[:], scattered.view[:])
+        assert_quantity_equals(result, scattered)
 
 
-def test_tile_gather_with_recv_quantity(
-    tile_quantity, scattered_quantities, communicator_list
+def test_cube_gather_with_recv_quantity(
+    cube_quantity, scattered_quantities, communicator_list
 ):
-    recv_quantity = copy.deepcopy(tile_quantity)
+    recv_quantity = copy.deepcopy(cube_quantity)
     recv_quantity.data[:] = -1
     for communicator, rank_quantity in reversed(
         list(zip(communicator_list, scattered_quantities))
@@ -285,18 +278,13 @@ def test_tile_gather_with_recv_quantity(
         else:
             result = communicator.gather(send_quantity=rank_quantity)
             assert result is None
-    assert recv_quantity.dims == tile_quantity.dims
-    assert recv_quantity.units == tile_quantity.units
-    assert recv_quantity.extent == tile_quantity.extent
-    tile_quantity.np.testing.assert_array_equal(
-        recv_quantity.view[:], tile_quantity.view[:]
-    )
+    assert_quantity_equals(recv_quantity, cube_quantity)
 
 
-def test_tile_scatter_state(
-    tile_quantity, scattered_quantities, communicator_list, time
+def test_cube_scatter_state(
+    cube_quantity, scattered_quantities, communicator_list, time
 ):
-    state = {"time": time, "air_temperature": tile_quantity}
+    state = {"time": time, "air_temperature": cube_quantity}
     result_list = []
     for communicator in communicator_list:
         if communicator.rank == 0:
@@ -306,16 +294,13 @@ def test_tile_scatter_state(
     for result_state, scattered in zip(result_list, scattered_quantities):
         assert result_state["time"] == time
         result = result_state["air_temperature"]
-        assert result.dims == scattered.dims
-        assert result.units == scattered.units
-        assert result.extent == scattered.extent
-        scattered.np.testing.assert_array_equal(result.view[:], scattered.view[:])
+        assert_quantity_equals(result, scattered)
 
 
-def test_tile_scatter_state_with_recv_state(
-    tile_quantity, scattered_quantities, communicator_list, time
+def test_cube_scatter_state_with_recv_state(
+    cube_quantity, scattered_quantities, communicator_list, time
 ):
-    tile_state = {"time": time, "air_temperature": tile_quantity}
+    tile_state = {"time": time, "air_temperature": cube_quantity}
     recv_quantities = copy.deepcopy(scattered_quantities)
     for q in recv_quantities:
         q.data[:] = 0.0
@@ -333,7 +318,4 @@ def test_tile_scatter_state_with_recv_state(
     for rank, (result, scattered) in enumerate(
         zip(recv_quantities, scattered_quantities)
     ):
-        assert result.dims == scattered.dims
-        assert result.units == scattered.units
-        assert result.extent == scattered.extent
-        scattered.np.testing.assert_array_equal(result.view[:], scattered.view[:])
+        assert_quantity_equals(result, scattered)